Validate key names in line with MDB's requirements
Redis, by default, allows pretty much any byte sequence to be a key
(including the empty string), up to (reputedly) 2^31 or so octets.

MDB is not so forgiving.  It limits keys to 511 bytes, and pitches a
not-so-tiny fit on attempts to manipulate a value whose key is the empty
string.  So, in the interests of sanity, I've gone through and found
(hopefully) every command which can take one or more keys as arguments, and
caused them all to return errors on attempts to specify an invalid key.