Don't open a transaction more than once
It turns out that MDB has a requirement that only one transaction may
be active at a time per thread.  This means that when we have a txn
open in, say, a walker, we can't then open another one to read the key
to check its expiry.  The effort involved in making this work properly
is significant, and involves refcounting and checking things and is
potentially a bug-ridden mess, but there aren't any other alternatives
I could think of (suggestions appreciated).

Now, at least, we won't get MDB_BAD_RSLOT errors on KEYS.